requirements_gpu.txt
h5py
numpy
scikit-learn
tensorflow-gpu
keras
opencv-python
pyttsx3